perf: add_interactions! improvements

+++ src/datatypes/interaction.jl
@@ -2,5 +2,5 @@ struct Interaction <: AbstractInteraction
     match_id::Int
     individual_id::Int
     other_ids::Vector{Int}
-    score::Float32
+    score::Float64
 end
